@@ -1,14 +1,16 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import { ConfigProvider } from "antd";
+import { createRoot } from "react-dom/client";
 import zh_CN from "antd/lib/locale-provider/zh_CN";
 
 import Router from "./routes";
 import "./styles/normalize.less";
 
-ReactDOM.render(
+const container = document.getElementById("vincent");
+const root = createRoot(container!);
+
+root.render(
   <ConfigProvider locale={zh_CN}>
     <Router />
-  </ConfigProvider>,
-  document.getElementById("vincent")
+  </ConfigProvider>
 );
